fixed warning on kaiser in drt_discret_utils.cpp
minor in drt_utils_maxent_basisfunctions.H




git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@18237 060e22d8-9d1c-0410-868a-e3a950d63d95

+++ src/drt_lib/drt_discret_utils.cpp
@@ -196,7 +196,7 @@ void DRT::Discretization::ComputeNullSpace(
     nullspace->resize(dimns*lrows,0.0);
   }
   // check if nullspace has correct size
-  else if(nullspace->size()!=(dimns*lrows) or numdf==0)
+  else if((ssize_t)(nullspace->size())!=(dimns*lrows) or numdf==0)
   {
     dserror("nullspace does not have correct size or numdf or dimns are zero");
   }
